Auto Fixes
-Dmytro
-Aidan

+++ Vimmy2024/src/main/java/frc/robot/subsystems/DriverDisplay.java
@@ -134,7 +134,7 @@ public class DriverDisplay extends SubsystemBase {
   public void periodic() {
     if(ArmSubsystem.hasNote) ControllerRumble.RumbleBothControllersBothSides(0.5);
     else ControllerRumble.RumbleBothControllersBothSides(0);
-    AutoSuccesfullShots.setInteger(AutoSequences.succesfulShots);
+    AutoSuccesfullShots.setInteger((int)AutoSequences.succesfulShots);
     switch ((int)DriverDisplay.gyroOrientation.getInteger(0)) {
       case 0:
         angleToAssign = 180;
